+++ src/front/js/component/navbar/searchBar/searchOptions.js
@@ -1,0 +1,22 @@
+import React from "react";
+
+export const SearchOptions = ({ searchCategory, setSearchCategory }) => {
+    return (
+        <div className="search-options">
+            <button
+                type="button"
+                className={`option-button ${searchCategory === 'recetas' ? 'active' : ''}`}
+                onClick={() => setSearchCategory("recetas")}
+            >
+                Recetas
+            </button>
+            <button
+                type="button"
+                className={`option-button ${searchCategory === 'usuarios' ? 'active' : ''}`}
+                onClick={() => setSearchCategory("usuarios")}
+            >
+                Usuarios
+            </button>
+        </div>
+    );
+};
