+++ src/front/js/layout.js
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import ScrollToTop from "./component/scrollToTop";
 import { BackendURL } from "./component/backendURL";
-import { Navbar } from "./component/navbar/navbar";
+import { Navbar } from "./component/navbar";
 import { Footer } from "./component/footer/index";
